Hide load order storage mechanism
All load order interaction is now done through the getters and setters.
There are a couple of changes in behaviour:

1. lo_fix_plugin_lists() does not remove invalid plugins, because
they're not loaded from files and can't be added in a setter, so the
only way to encounter them is if a plugin is deleted, which would then
trigger the load order to be reloaded anyway.
2. lo_set_load_order() does not add any installed plugins not given to
the end of the load order, because this will be done on the next load
anyway.