Make game handle data storage thread-safe
For #2. Each game handle uses a separate key-value pair in a
static unordered map for data storage. The storage is thread-local
and will be created if missing for a game handle.

Dynamic memory management was removed to avoid leaks because
thread-local storage is deallocated when the thread is destroyed.
The use of an internal C++ string array is a little messy, but
necessary as C++ strings cannot be cast to C strings. The
const_cast() is used because the API doesn't output const strings
(though that's bad design).

There aren't any tests to check if the mutex is doing its job in
preventing data races, because I don't know how to test that.