Update Fallout 4 load order mechanism used
Fallout v1.5 introduces a new load order mechanism that lists all
installed plugins in their load order in plugins.txt, and prefixes
active plugins with an asterisk '*'.

This method has been named the asterisk load order method in
libloadorder, and this commit implements support for it, and adds
a new LIBLO_METHOD_ASTERISK constant that is used for Fallout 4.