Replace use of 'as' keyword
The FFI constant values are unchanged, this clarifies what they are.
There was an unnecessary pointer cast in the FFI layer. Some test code
assumed that an integer cast would never truncate, but it's better to
check that.

There's nothing really wrong with it being used to rename a type, but I
think the code's a little clearer without it.