Rework ability to load external plugins
It turns out that the Microsoft Store's Fallout 4 doesn't just load the
DLC plugins from their data directories, but will also load other
plugins in those directories, so instead of specifying the plugin paths,
specify the directory paths to scan for plugins.

This also reorders the default list of external directories to match the
order in which the game checks them.