Add support for Starfield
This adds support for override plugins, which do not use up a mod index
when active.

While plugins.txt and Starfield.ccc are disabled by default, it turns
out that it's trivial to enable by renaming
BlueprintShips-Starfield.esm.

The only thing left unimplemented is support for reading the
Starfield_<language>.INI file, because detecting the appropriate
language to use is complex enough that it'll be implemented separately.