Fix inconsistent save() and load() when plugins.txt is ignored
If Fallout 4 or Starfield are ignoring their plugins.txt then writing
plugins.txt is harmless but ineffective, as the saved load order won't
be seen when the load order is next loaded.

As such, if plugins.txt is ignored, write the load order by modifying
plugin modification timestamps, as if the load order was
timestamp-based. That way, the load order saved by save() will agree
with the load order read by load().