Fix paths used for OpenMW on Linux
Also support non-default global configurations on Windows and Linux.

OpenMW's approach to configuration is quite complicated, so has been
moved to a separate module. The approach taken makes creating a
GameSettings object a bit inefficient as the same config is loaded
multiple times, but changing that would require a wider refactoring of
GameSettings creation that would involve code used for other games.

One thing that tripped me up while making these changes is that
Ini::is_empty() checks if there are any sections, but Ini::new() creates
an object with an anonymous/general section, so a new ini object is not
empty, even though it contains no fields.