Suppress audit suggestions for crates not used on target platforms
I only test libloadorder for x86-64 Windows and x86-64 Linux in CI. It is
also used for x86 Windows builds in libloot, but that'll almost
certainly drop 32-bit builds for its next release.

Ideally there would be a way to tell cargo-vet to ignore dependencies
that aren't relevant to a given set of targets, as having them exempt
like this could hide the need to audit them if they do become relevant
to those targets in the future, but without the supression `cargo vet
suggest`'s output is just too noisy.